# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the login endpoint",
        "contact": {
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/quests": {
            "get": {
                "tags": [
                    "Quests"
                ],
                "summary": "Display a listing of the Quests",
                "operationId": "getQuestsList",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token for authentication",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer {token}"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quests retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Quests"
                ],
                "summary": "Store a newly created Quest in storage",
                "operationId": "storeQuest",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "difficulty",
                                    "recurrence",
                                    "cooldown",
                                    "claim_time",
                                    "condition",
                                    "reward",
                                    "module",
                                    "sprint",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Quest name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Quest description"
                                    },
                                    "difficulty": {
                                        "type": "string",
                                        "example": "easy"
                                    },
                                    "recurrence": {
                                        "type": "string",
                                        "example": "daily"
                                    },
                                    "cooldown": {
                                        "type": "integer",
                                        "example": 24
                                    },
                                    "claim_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-05-18T00:00:00Z"
                                    },
                                    "condition": {
                                        "type": "string",
                                        "example": "Complete 3 tasks"
                                    },
                                    "reward": {
                                        "type": "string",
                                        "example": "100 points"
                                    },
                                    "module": {
                                        "type": "string",
                                        "example": "Module A"
                                    },
                                    "sprint": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "status": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Quest saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Quest name"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Quest description"
                                                },
                                                "difficulty": {
                                                    "type": "string",
                                                    "example": "easy"
                                                },
                                                "recurrence": {
                                                    "type": "string",
                                                    "example": "daily"
                                                },
                                                "cooldown": {
                                                    "type": "integer",
                                                    "example": 24
                                                },
                                                "claim_time": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-05-18T00:00:00Z"
                                                },
                                                "condition": {
                                                    "type": "string",
                                                    "example": "Complete 3 tasks"
                                                },
                                                "reward": {
                                                    "type": "string",
                                                    "example": "100 points"
                                                },
                                                "module": {
                                                    "type": "string",
                                                    "example": "Module A"
                                                },
                                                "sprint": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "status": {
                                                    "type": "integer",
                                                    "example": "1"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Quest saved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bad request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/quests/{id}": {
            "delete": {
                "tags": [
                    "Quests"
                ],
                "summary": "Remove the specified Quests from storage",
                "description": "Deletes a quests by its ID",
                "operationId": "destroyQuests",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the quests to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quests deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Quests deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Quests not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Quests not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}